requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
flasgger==0.9.7.1
Flask==3.0.0
gunicorn==21.2.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
MarkupSafe==2.1.3
mistune==3.0.2
packaging==23.2
polygon-api-client==1.13.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
rpds-py==0.13.1
six==1.16.0
urllib3==1.26.18
websockets==12.0
Werkzeug==3.0.1
